função gameOver passada para Game.js
